Bug fix, visibility changes
Fix a bug where we could reset the algorithm before any iterations.
Changed tree display from using a counter to stop drawing the current
node before iterations (which didnt work after reset or undo) to
checking whether we can undo or not (if we cant undo we must be at
begining). Visibility changes for the search creators